\docType{data}
\name{AMF_giovannetti}
\alias{AMF_giovannetti}
\title{AMF inoculation potential}
\format{a dataframe of 240 rws with the following 2 variables \describe{

  \item{species}{name of plant species}
  \item{support.AMF}{score which assess the "potentiality for the species to host AMF"}

}}
\description{
AMF inoculation potential
}
\examples{
\dontrun{
#' data(AMF_giovannetti)
}
}
\references{
Giovannetti E., Avio L., Sbrana C. (personal communication)
}
\seealso{
\code{help(retrieve_amf)}
}
\keyword{datasets}

