\docType{data}
\name{biolflor_lu}
\alias{biolflor_lu}
\title{biolflor_lu}
\format{A data frame with 3688 observations on the following 8 variables.
\describe{
\item{V1}{name of plant species}
\item{V2}{url of the Biolflor webpage containing traits data for the species}
}}
\description{
A lookup dataframe for retrieving data from the Biolflor
website
}
\examples{
\dontrun{
data(biolflor_lu)
}
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}
\references{
Please use the follow citation ay time you use data derived
from Biolflor:

Klotz, S., Kühn, I., Durka, W. (eds) (2002). BIOLFLOR - Eine Datenbank 
zu biologisch-ökologischen Merkmalen zur Flora von Deutschland. 
Schriftenreihe für Vegetationskunde *38*: 1-333.  (Bundesamt für. Bonn, 
Bundesamt für Naturschutz)
}
\keyword{dataframe}

