\name{column_conversion}
\alias{column_conversion}
\title{column_conversion}
\usage{
column_conversion(DF)
}
\arguments{
  \item{DF}{a intermediate dataframe built by \code{tr8}}
}
\value{
a dataframe with shortened col names
}
\description{
Shortens up columns' names in the intermidiate steps of
tr8() retrieved data
}
\details{
This function is internally used by \code{tr8()}.
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}

