\docType{data}
\name{list_of_traits_Biolflor}
\alias{list_of_traits_Biolflor}
\title{list_of_traits_Biolflor}
\format{A vector of plant traits}
\description{
a vector containing traits that can be downloaded from
Biolflor
}
\examples{
## have a look at the first rows of the dataset
head(list_of_traits_Biolflor)

}
\author{
Gionata Bocci <boccigionata@gmail.com>
}
\references{
Please use the follow citation ay time you use data derived
from Biolflor:

Klotz, S., Kühn, I., Durka, W. (eds) (2002). BIOLFLOR - Eine Datenbank 
zu biologisch-ökologischen Merkmalen zur Flora von Deutschland. 
Schriftenreihe für Vegetationskunde *38*: 1-333.  (Bundesamt für. Bonn, 
Bundesamt für Naturschutz)
}
\keyword{list_of_traits_Biolflor}

