\docType{data}
\name{myco}
\alias{myco}
\title{myco}
\format{A data frame with 2970 observations on the following 2 variables.
\describe{
\item{species}{name of plant species}
\item{Myco_infection}{score of AMF infection}
 }}
\description{
A lookup dataframe for checking species names on the
Biolflor website confronting them with the \code{tnrs}
function
}
\examples{
\dontrun{
#data(myco)
}
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}
\references{

Akhmetzhanova, A.A, Soudzilovskaia, N.A., Onipchenko, V.G.,
Cornwell, W. K., Agafonov, V. A., Selivanov, I.A., and Cornelissen, 
J. H. C. (2012): A rediscovered treasure: mycorrhizal intensity database for
3000 vascular plants species across the former Soviet
Union. Ecology 93:689. \url{http://esapubs.org/Archive/ecol/E093/059/default.htm}
}
\keyword{dataframe}

