
\docType{data}
\name{pignatti}
\alias{pignatti}
\title{pignatti}
\format{Contains the following traits data about Italian Flora species\describe{
 \item{Specie.Pignatti }{: species name in the original dataset }
 \item{numero          }{: numeric code }
 \item{codice          }{: numeric code }
 \item{nome.scientifico}{: scientific name with authors }
 \item{forma_biologica }{: life form}
 \item{corotipo        }{: distribution of species }
 \item{L               }{: Ellenberg value for light}
 \item{T               }{: Ellenberg value for temperature}
 \item{C               }{: Ellenberg value for continentality}
 \item{U               }{: Ellenberg value for soil humidity }
 \item{R               }{: Ellenberg value for soil pH}
 \item{N               }{: Ellenberg value for Nutrients in the soil }
 \item{S               }{: Ellenberg value for soil salinity}
 \item{Name.tnrs       }{: species name according to tnrs }
}}
\description{
\code{pignatti} is a dataframe containing traits data for
italian species
}
\examples{
## inspect the first rows of the dataset
head(pignatti)
}
\references{
Pignatti, S., Menegoni, P., Pietrosanti, S., 2005,
Biondicazione attraverso le piante vascolari. Valori di
indicazione secondo Ellenberg (Zeigerwerte) per le specie
della Flora d'Italia. Braun-Blanquetia 39, Camerino, pp.
97.
}
\keyword{datasets}

