\name{pignatti_f}
\alias{pignatti_f}
\title{Extracts ellenberg values for the Italia Flora as provided by Pignatti et al. (2005)}
\usage{
pignatti_f(species, TRAITS)
}
\arguments{
  \item{species}{a vector containing plant species names}

  \item{TRAITS}{a vector containing the traits to be
  downloaded (used as a check for tr8_gui() created
  variables)}
}
\value{
a data frame
}
\description{
This function is not ment to be used by the final user;
it's used by the wrapper function \code{tr8()}
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}
\references{
Pignatti, S., Menegoni, P., Pietrosanti, S., 2005,
Biondicazione attraverso le piante vascolari. Valori di
indicazione secondo Ellenberg (Zeigerwerte) per le specie
della Flora d'Italia. Braun-Blanquetia 39, Camerino, pp.
97.
}

