\name{tr8_config}
\alias{tr8_config}
\title{\code{tr8_config} a GUI to configure TR8 package.}
\usage{
tr8_config()
}
\value{
Returns a list containing the selected
traits for each source of information
}
\description{
This function will make a simple GUI appear which
contains a tab for each trait database: the user can choose
which traits should be downloaded by the \code{tr8}
function.
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}
\seealso{
tr8()
}

