\name{tr8_setup}
\alias{tr8_setup}

\title{
  A function to download the most recent url-tables for some traitbases.
}
\description{
  For some traitbases tr8 uses lookup-tables which contain URLs of the
  species of interest; some of the traitbases queried by tr8 are
  uploaded from time to time thus the function may be unable to retrieve
  traits data for the most recently uploaded species.
  This function allows the user to refresh some of these lookup tables.
  \bold{BEWARE}: this function takes a long time to run (nearly 20
  minutes are required - on a fast connection - for retrieving lookup
  tables for BiolFlor). 
}
\usage{
tr8_setup()
}

\details{
  At the moment BiolFlor and Ecoflora lookup tables are re-generated by this function.
}
\value{
  The values does not return anythis useful for the user; it's used for
  its side effects, i.e. it stores the retrieved lookup table in one
  directory for later use from the \code{tr8} function.
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}

\section{Warning }{
  The function takes a lot to run; I suggest to run it just at your
  first installation of the \code{TR8} package.
} 

\examples{
\dontrun{
tr8_setup()
}
}

\keyword{TR8 }
\keyword{tr8 }
