\docType{data}
\name{traits_special_Biolflor}
\alias{traits_special_Biolflor}
\title{traits_special_Biolflor}
\format{A vector of plant traits}
\description{
a vector containing traits that can be downloaded from
Biolflor for which some special Xpaht rules must be applied
}
\examples{
traits_special_Biolflor
}
\author{
Gionata Bocci <boccigionata@gmail.com>
}
\references{
Please use the follow citation ay time you use data derived
from Biolflor: BIOLFLOR - Eine Datenbank zu
biologisch-ökologischen Merkmalen zur Flora von
Deutschland. Schriftenreihe für Vegetationskunde 38: 1-333.
(Bundesamt für. Bonn, Bundesamt für Naturschutz)
}
\keyword{traits_special_Biolflor}

